Out the file using nameSpace
