+++ src/app.ts
@@ -1,29 +1,8 @@
-// Drag and Drop Interfaces
-interface Draggable{
-  dragStartHandler(event: DragEvent): void
-  dragEndHandler(event: DragEvent): void
-}
-
-interface DragTarget {
-  dragOverHandler(event:DragEvent): void;
-  dropHandler(event:DragEvent): void;
-  dragLeaveHandler(event:DragEvent): void;
-}
+/// <reference path="drag-drop-interfaces.ts" />
+/// <reference path="project-model.ts" />
 
 
- // Project Type
-enum ProjectStatus {active, finished}
-
-class Project {
-  constructor(
-    public id: string,
-    public title: string,
-    public description: string,
-    public people: number,
-    public status: ProjectStatus){
-
-  }
-}
+namespace App {
 
 type Listener<T> = (items: T[]) => void
 
@@ -361,8 +340,12 @@ class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
   }
 
 
-}
+  }
 
-const prjInput = new ProjectInput();
-const activePrjList = new ProjectList('active');
-const finishedPrjList = new ProjectList('finished');
+  new ProjectInput();
+  new ProjectList('active');
+  new ProjectList('finished');
+
+
+  }
+
