+++ src/project-model.ts
@@ -1,0 +1,12 @@
+// Project Type
+enum ProjectStatus {active, finished}
+
+class Project {
+    constructor(
+    public id: string,
+    public title: string,
+    public description: string,
+    public people: number,
+    public status: ProjectStatus){
+    }
+}
